refactor(project) Using the org.elasticsearch.client.rest client
Refactored the Client class to submit request and get responses
using the new RestClient.
Parse responses using Jackson 2.8 to get the results in POJOs
Removed dependencies on elasticsearch 2.3.x
Added support for async operations based on the new client
Renamed some annotations and types of the to-be-generated classes
Search operations are triggered from a delegate that is instanciated
when a query is being constructed
CodecRegistry and Codecs refactoring: the codec registry is specific
to each document manager and provides with Document codec(s) to
read/write documents, and a DocumentSearchCodec to convert a query
into a JSON document sent by the rest client. The Java<->JSON
conversion is performed by Jackson using custom JsonSerializers
and JsonDeserializers.

Using docker-maven-plugin to start/stop Elasticsearch during
integration tests, but in a separate profile because travis-ci
cannot start Elasticsearch in a Docker container yet.